+++ test/cljdub_dsp/io_test.clj
@@ -14,10 +14,27 @@
            1))
     (is (every? #(= % 0) (samples seg)))))
 
+(deftest from-bytes-test
+  (let [byte-string (bins/bytes audio-data)
+        seg (from-bytes audio-data
+                        :sample-width audio-sample-width
+                        :sample-rate audio-frame-rate
+                        :channels audio-channels)]
+    (is (= (raw-data seg)
+           byte-string))
+    (is (= (sample-width seg)
+           audio-sample-width))
+    (is (= (frame-rate seg)
+           audio-frame-rate))
+    (is (= (channels seg)
+           audio-channels))))
+
 (deftest from-byte-string-test
   (let [byte-string (bins/bytes audio-data)
-        seg (from-byte-string byte-string audio-sample-width
-                              audio-frame-rate audio-channels)]
+        seg (from-byte-string byte-string
+                              :sample-width audio-sample-width
+                              :sample-rate audio-frame-rate
+                              :channels audio-channels)]
     (is (= (raw-data seg)
            byte-string))
     (is (= (sample-width seg)